+++ solores/src/idl_format/bincode/mod.rs
@@ -2,31 +2,26 @@ use serde::Deserialize;
 use toml::{map::Map, Value};
 
 use crate::write_cargotoml::{
-    DependencyValue, FeaturesDependencyValue, OptionalDependencyValue, BORSH_CRATE, BYTEMUCK_CRATE,
-    NUM_DERIVE_CRATE, NUM_TRAITS_CRATE, SERDE_CRATE, SOLANA_PROGRAM_CRATE, THISERROR_CRATE,
+    DependencyValue, NUM_DERIVE_CRATE, NUM_TRAITS_CRATE, SERDE_CRATE, SOLANA_PROGRAM_CRATE,
+    THISERROR_CRATE,
 };
 
 use super::{IdlCodegenModule, IdlFormat};
 
 use self::{
-    accounts::AccountsCodegenModule,
     errors::{ErrorEnumVariant, ErrorsCodegenModule},
     instructions::{IxCodegenModule, NamedInstruction},
-    typedefs::{NamedType, TypedefsCodegenModule},
 };
 
-pub mod accounts;
 pub mod errors;
 pub mod instructions;
 pub mod typedefs;
 
 #[derive(Deserialize)]
-pub struct ShankIdl {
+pub struct BincodeIdl {
     pub name: String,
     pub version: String,
     pub metadata: Metadata,
-    pub accounts: Option<Vec<NamedType>>,
-    pub types: Option<Vec<NamedType>>,
     pub instructions: Option<Vec<NamedInstruction>>,
     pub errors: Option<Vec<ErrorEnumVariant>>,
 }
@@ -37,7 +32,7 @@ pub struct Metadata {
     pub origin: String,
 }
 
-impl IdlFormat for ShankIdl {
+impl IdlFormat for BincodeIdl {
     fn program_name(&self) -> &str {
         &self.name
     }
@@ -51,23 +46,11 @@ impl IdlFormat for ShankIdl {
     }
 
     fn is_correct_idl_format(&self) -> bool {
-        self.metadata.origin == "shank"
+        self.metadata.origin == "bincode"
     }
 
-    fn modules<'me>(&'me self, args: &'me crate::Args) -> Vec<Box<dyn IdlCodegenModule + 'me>> {
+    fn modules<'me>(&'me self, _args: &'me crate::Args) -> Vec<Box<dyn IdlCodegenModule + 'me>> {
         let mut res: Vec<Box<dyn IdlCodegenModule + 'me>> = Vec::new();
-        if let Some(v) = &self.accounts {
-            res.push(Box::new(AccountsCodegenModule {
-                cli_args: args,
-                named_types: v,
-            }));
-        }
-        if let Some(v) = &self.r#types {
-            res.push(Box::new(TypedefsCodegenModule {
-                cli_args: args,
-                named_types: v,
-            }));
-        }
         if let Some(v) = &self.instructions {
             res.push(Box::new(IxCodegenModule {
                 program_name: self.program_name(),
@@ -85,25 +68,11 @@ impl IdlFormat for ShankIdl {
 
     fn dependencies(&self, args: &crate::Args) -> Map<String, Value> {
         let mut map = Map::new();
-        map.insert(BORSH_CRATE.into(), DependencyValue(&args.borsh_vers).into());
-        if !args.zero_copy.is_empty() {
-            map.insert(
-                BYTEMUCK_CRATE.into(),
-                FeaturesDependencyValue {
-                    dependency: DependencyValue(&args.bytemuck_vers),
-                    features: vec!["derive".into()],
-                }
-                .into(),
-            );
-        }
         map.insert(
             SOLANA_PROGRAM_CRATE.into(),
             DependencyValue(&args.solana_program_vers).into(),
         );
-        map.insert(
-            SERDE_CRATE.into(),
-            OptionalDependencyValue(DependencyValue(&args.serde_vers)).into(),
-        );
+        map.insert(SERDE_CRATE.into(), DependencyValue(&args.serde_vers).into());
         if self.errors.is_some() {
             map.insert(
                 THISERROR_CRATE.into(),